sql/postgres/requirements.txt
Faker==18.7.0
psycopg2-binary==2.9.6
python-dateutil==2.8.2
six==1.16.0
SQLAlchemy==2.0.13
typing_extensions==4.5.0
